design: 사이드바 디자인 변경
