+++ client/src/features/albumRegister/model/useAlbumForm.ts
@@ -17,8 +17,6 @@ const SONG_FIELDS = [...REQUIRED_SONG_FIELDS, 'lyrics'] as const;
 
 const ALBUM_FIELDS = ['title', 'artist', 'albumTag', 'releaseDate'] as const;
 
-//const API_URL = 'http://localhost:3000/api/admin/album';
-
 function validateForm(formData: FormData, fields: readonly string[]): boolean {
   return fields.every((field) => {
     const value = formData.get(field);
